[Refactor] chat diary refactor
