+++ frontend/src/constants/constants.ts
@@ -1,4 +1,4 @@
-export const LOAD_DAIRY = 5; // 내 일기 목록 무한 스크롤로 가져 올 떄 한번에 불러올 개수, 다른 이름이나 방식 있으면 수정예정
+export const LOAD_DAIRY = 2; // 내 일기 목록 무한 스크롤로 가져 올 떄 한번에 불러올 개수, 다른 이름이나 방식 있으면 수정예정
 export const IMAGE_LIMIT = 5;
 export const TAG_LIMIT = 5;
 export const DEFAULT_TIMEOUT = 30000;
