+++ frontend/src/hooks/chat/useChatList.ts
@@ -11,6 +11,7 @@ export const useChatList = () => {
         return;
       }
       setChatList(data);
+      console.log(data);
     });
   }, []);
